requirements.txt
chainlit
langchain==0.0.304
python-dotenv
google-generativeai
ipykernel
chromadb
pypdf
sentence-transformers
openai
google-search-results
pandas
msal